Rename RequestFactoryParser to RequestFactory.Builder.
It's not perfectly a builder, but it fits the pattern of embedding the builder
or factory within a type.

Also rename Response.isSuccess() to Response.isSuccessful().

Also rename RequestAction to ParameterAction. These are only used for
request parameters.

Also move all utility methods into the same class, Utils.java.
