+++ retrofit/src/main/java/retrofit2/Callback.java
@@ -33,7 +33,7 @@ public interface Callback<T> {
    * Invoked for a received HTTP response.
    * <p>
    * Note: An HTTP response may still indicate an application-level failure such as a 404 or 500.
-   * Call {@link Response#isSuccess()} to determine if the response indicates success.
+   * Call {@link Response#isSuccessful()} to determine if the response indicates success.
    */
   void onResponse(Call<T> call, Response<T> response);
